openai_realtime_voice_agent/pyproject.toml
[tool.poetry]
name = "openai-realtime-voice-agent"
version = "0.1.0"
description = "Home Assistant addon for OpenAI Realtime API voice control with WebSocket support using Pipecat"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "app", from = "."}]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
pipecat-ai = {extras = ["mcp", "openai"], version = "*"}
python-dotenv = ">=1.0.0"
websockets = ">=12.0"  # Required for custom WebSocket transport

[tool.poetry.scripts]
start = "app.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
